requirements.txt
# Django packages
django==2.1.2
djangorestframework==3.9.0
django-cors-headers==2.2.0
django-rest-auth==0.9.3
Pillow==5.1.0

# Sql packages
psycopg2==2.7.5

# Utility packages
python-decouple==3.1
coreapi==2.3.3
ipython==6.5.0

# Linting packages
flake8==3.5.0
flake8-mypy==17.3.3
autopep8==1.3.5
